Add support for Azure Identities
Extend the Azure Blob client with support for Azure Identity credentials
[1][2]. Since these credentials are dynamically retrieved from the
metadata service of the Virtual Machine or AKS pod, use this
method when no credentials have been provided.

Note that currently we perform an HTTP request to retrieve the token
instead of using a dedicated client, and we do not cache the token until
it expires, i.e., we retrieve a new token for each Azure Blob request.

[1] https://docs.microsoft.com/en-us/azure/active-directory/managed-identities-azure-resources/how-to-use-vm-token
[2] https://docs.microsoft.com/en-us/azure/aks/use-azure-ad-pod-identity

Signed-off-by: Kostis Lolos <klolos@arrikto.com>
Reviewed-by: Nikos Dragazis <ndragazis@arrikto.com>